+++ sdk/keyvault/keyvault-secrets/samples/typescript/helloWorld.ts
@@ -33,7 +33,8 @@ async function main(): Promise<void> {
   console.log("updated secret: ", updatedSecret);
 
   // Delete the secret
-  await client.deleteSecret(secretName);
+  const deletePoller = await client.beginDeleteSecret(secretName);
+  await deletePoller.pollUntilDone();
 }
 
 main().catch((err) => {