+++ sdk/keyvault/keyvault-secrets/samples/typescript/listOperations.ts
@@ -44,8 +44,11 @@ async function main(): Promise<void> {
     console.log("secret version: ", secret);
   }
 
-  await client.deleteSecret(bankAccountSecretName);
-  await client.deleteSecret(storageAccountSecretName);
+  let deletePoller = await client.beginDeleteSecret(bankAccountSecretName);
+  await deletePoller.pollUntilDone();
+
+  deletePoller = await client.beginDeleteSecret(storageAccountSecretName);
+  await deletePoller.pollUntilDone();
 }
 
 main().catch((err) => {